[IMP] account_edi: merge account_edi_extended
TASK ID: 2457873
