+++ addons/account_edi/models/account_edi_document.py
@@ -2,11 +2,12 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from odoo import models, fields, api
-from odoo.addons.account_edi_extended.models.account_edi_document import DEFAULT_BLOCKING_LEVEL
 from psycopg2 import OperationalError
 import logging
 
 _logger = logging.getLogger(__name__)
+
+DEFAULT_BLOCKING_LEVEL = 'error'
 
 
 class AccountEdiDocument(models.Model):
@@ -19,6 +20,12 @@ class AccountEdiDocument(models.Model):
     attachment_id = fields.Many2one('ir.attachment', help='The file generated by edi_format_id when the invoice is posted (and this document is processed).')
     state = fields.Selection([('to_send', 'To Send'), ('sent', 'Sent'), ('to_cancel', 'To Cancel'), ('cancelled', 'Cancelled')])
     error = fields.Html(help='The text of the last error that happened during Electronic Invoice operation.')
+    blocking_level = fields.Selection(
+        selection=[('info', 'Info'), ('warning', 'Warning'), ('error', 'Error')],
+        help="Blocks the document current operation depending on the error severity :\n"
+        "  * Info: the document is not blocked and everything is working as it should.\n"
+        "  * Warning : there is an error that doesn't prevent the current Electronic Invoicing operation to succeed.\n"
+        "  * Error : there is an error that blocks the current Electronic Invoicing operation.")
 
     # == Not stored fields ==
     name = fields.Char(related='attachment_id.name')
@@ -31,14 +38,6 @@ class AccountEdiDocument(models.Model):
             'Only one edi document by move by format',
         ),
     ]
-
-    def write(self, vals):
-        ''' If account_edi_extended is not installed, a default behaviour is used instead.
-        '''
-        if 'blocking_level' in vals and 'blocking_level' not in self.env['account.edi.document']._fields:
-            vals.pop('blocking_level')
-
-        return super().write(vals)
 
     def _prepare_jobs(self):
         """Creates a list of jobs to be performed by '_process_job' for the documents in self.
@@ -53,10 +52,7 @@ class AccountEdiDocument(models.Model):
 
         # Classify jobs by (edi_format, edi_doc.state, doc_type, move.company_id, custom_key)
         to_process = {}
-        if 'blocking_level' in self.env['account.edi.document']._fields:
-            documents = self.filtered(lambda d: d.state in ('to_send', 'to_cancel') and d.blocking_level != 'error')
-        else:
-            documents = self.filtered(lambda d: d.state in ('to_send', 'to_cancel'))
+        documents = self.filtered(lambda d: d.state in ('to_send', 'to_cancel') and d.blocking_level != 'error')
         for edi_doc in documents:
             move = edi_doc.move_id
             edi_format = edi_doc.edi_format_id
@@ -87,36 +83,6 @@ class AccountEdiDocument(models.Model):
             if batch:
                 target.append((batch, doc_type))
         return invoices + payments
-
-    @api.model
-    def _convert_to_old_jobs_format(self, jobs):
-        """ See '_prepare_jobs' :
-        Old format : ((edi_format, state, doc_type, company_id), documents)
-        Since edi_format, state and company_id can be deduced from documents, this is redundant and more prone to unexpected behaviours.
-        New format : (doc_type, documents).
-
-        However, for backward compatibility of 'process_jobs', we need a way to convert back to the old format.
-        """
-        return [(
-            (documents.edi_format_id, documents[0].state, doc_type, documents.move_id.company_id),
-            documents
-        ) for documents, doc_type in jobs]
-
-    @api.model
-    def _process_jobs(self, to_process):
-        """ Deprecated, use _process_job instead.
-
-        :param to_process: A list of tuples (key, documents)
-        * key:             A tuple (edi_format_id, state, doc_type, company_id)
-        ** edi_format_id:  The format to perform the operation with
-        ** state:          The state of the documents of this job
-        ** doc_type:       Are the moves of this job invoice or payments ?
-        ** company_id:     The company the moves belong to
-        * documents:       The documents related to this job. If edi_format_id does not support batch, length is one
-        """
-        for key, documents in to_process:
-            dummy, dummy, doc_type, dummy = key
-            self._process_job(documents, doc_type)
 
     @api.model
     def _process_job(self, documents, doc_type):
@@ -220,7 +186,8 @@ class AccountEdiDocument(models.Model):
         """ Post and cancel all the documents that don't need a web service.
         """
         jobs = self.filtered(lambda d: not d.edi_format_id._needs_web_services())._prepare_jobs()
-        self._process_jobs(self._convert_to_old_jobs_format(jobs))
+        for documents, doc_type in jobs:
+            self._process_job(documents, doc_type)
 
     def _process_documents_web_services(self, job_count=None, with_commit=True):
         """ Post and cancel all the documents that need a web service. This is called by CRON.
