+++ addons/account_edi/models/account_payment.py
@@ -9,3 +9,7 @@ class AccountPayment(models.Model):
 
     def action_process_edi_web_services(self):
         return self.move_id.action_process_edi_web_services()
+
+    def action_retry_edi_documents_error(self):
+        self.ensure_one()
+        return self.move_id.action_retry_edi_documents_error()
