+++ addons/account_edi/views/account_move_views.xml
@@ -16,13 +16,21 @@
             <field name="arch" type="xml">
                 <xpath expr="//button[@name='button_cancel']" position="after">
                     <field name="edi_show_cancel_button" invisible="1"/>
+                    <field name="edi_show_abandon_cancel_button" invisible="1"/>
                     <button name="button_cancel_posted_moves"
                             string="Request EDI Cancellation"
                             type="object"
                             groups="account.group_account_invoice"
                             attrs="{'invisible' : [('edi_show_cancel_button', '=', False)]}"/>
+                    <button name="button_abandon_cancel_posted_posted_moves"
+                            string="Call off EDI Cancellation"
+                            type="object"
+                            groups="account.group_account_invoice"
+                            attrs="{'invisible' : [('edi_show_abandon_cancel_button', '=', False)]}"/>
                 </xpath>
                 <xpath expr="//header" position="after">
+                    <field name="edi_blocking_level" invisible="1" />
+                    <field name="edi_error_count" invisible="1" />
                     <div class="alert alert-info" role="alert" style="margin-bottom:0px;"
                         attrs="{'invisible': ['|', ('edi_web_services_to_process', 'in', ['', False]), ('state', '=', 'draft')]}">
                          <div>The invoice will be sent asynchronously to :
@@ -30,11 +38,29 @@
                          </div>
                          <button name="action_process_edi_web_services" type="object" class="oe_link" string="Send now" /> 
                     </div>
+                    <div class="alert alert-danger" role="alert" style="margin-bottom:0px;"
+                        attrs="{'invisible': ['|', ('edi_error_count', '=', 0), ('edi_blocking_level', '!=', 'error')]}">
+                        <div class="o_row">
+                            <field name="edi_error_message" />
+                            <button name="%(account_edi.action_open_edi_documents)d" string="⇒ See errors" type="action" class="oe_link" attrs="{'invisible': [('edi_error_count', '=', 1)]}" /> 
+                            <button name="action_retry_edi_documents_error" type="object" class="oe_link oe_inline" string="Retry" />
+                        </div>
+                    </div>
                     <div class="alert alert-warning" role="alert" style="margin-bottom:0px;"
-                         attrs="{'invisible': [('edi_error_count', '=', 0)]}">
-                         <field name="edi_error_count"/>
-                         Electronic invoicing error(s)
-                         <button name="%(account_edi.action_open_edi_documents)d" type="action" class="oe_link oe_inline" string="⇒ See errors" /> 
+                        attrs="{'invisible': ['|', ('edi_error_count', '=', 0), ('edi_blocking_level', '!=', 'warning')]}">
+                        <div class="o_row">
+                            <field name="edi_error_message" />
+                            <button name="%(account_edi.action_open_edi_documents)d" string="⇒ See errors" type="action" class="oe_link" attrs="{'invisible': [('edi_error_count', '=', 1)]}" /> 
+                            <button name="action_retry_edi_documents_error" type="object" class="oe_link oe_inline" string="Retry" />
+                        </div>
+                    </div>
+                    <div class="alert alert-info" role="alert" style="margin-bottom:0px;"
+                        attrs="{'invisible': ['|', ('edi_error_count', '=', 0), ('edi_blocking_level', '!=', 'info')]}">
+                        <div class="o_row">
+                            <field name="edi_error_message" />
+                            <button name="%(account_edi.action_open_edi_documents)d" string="⇒ See errors" type="action" class="oe_link" attrs="{'invisible': [('edi_error_count', '=', 1)]}" /> 
+                            <button name="action_retry_edi_documents_error" type="object" class="oe_link oe_inline" string="Retry" />
+                        </div>
                     </div>
                 </xpath>
                 <xpath expr="//div[@name='journal_div']" position="after">
