+++ addons/account_edi/views/account_payment_views.xml
@@ -15,6 +15,8 @@
             <field name="inherit_id" ref="account.view_account_payment_form" />
             <field name="arch" type="xml">
                 <xpath expr="//header" position="after">
+                    <field name="edi_blocking_level" invisible="1" />
+                    <field name="edi_error_count" invisible="1" />
                     <div class="alert alert-info" role="alert" style="margin-bottom:0px;"
                          attrs="{'invisible': ['|', ('edi_web_services_to_process', 'in', ['', False]), ('state', '=', 'draft')]}">
                          <div>The payment will be sent asynchronously to :
@@ -22,11 +24,27 @@
                          </div>
                          <button name="action_process_edi_web_services" type="object" class="oe_link" string="Send now"/>
                     </div>
+                    <div class="alert alert-danger" role="alert" style="margin-bottom:0px;"
+                        attrs="{'invisible': ['|', ('edi_error_count', '=', 0), ('edi_blocking_level', '!=', 'error')]}">
+                        <div class="o_row">
+                            <field name="edi_error_message" />
+                            <button name="%(account_edi.action_open_edi_documents)d" string="⇒ See errors" type="action" class="oe_link" attrs="{'invisible': [('edi_error_count', '=', 1)]}" /> 
+                            <button name="action_retry_edi_documents_error" type="object" class="oe_link oe_inline" string="Retry" />
+                        </div>
+                    </div>
                     <div class="alert alert-warning" role="alert" style="margin-bottom:0px;"
-                         attrs="{'invisible': [('edi_error_count', '=', 0)]}">
-                         <field name="edi_error_count"/>
-                         Electronic invoicing error(s)
-                         <button name="%(account_edi.action_open_payment_edi_documents)d" type="action" class="oe_link oe_inline" string="⇒ See errors" /> 
+                        attrs="{'invisible': ['|', ('edi_error_count', '=', 0), ('edi_blocking_level', '!=', 'warning')]}">
+                        <div class="o_row">
+                            <field name="edi_error_message" />
+                            <button name="%(account_edi.action_open_edi_documents)d" string="⇒ See errors" type="action" class="oe_link" attrs="{'invisible': [('edi_error_count', '=', 1)]}" /> 
+                        </div>
+                    </div>
+                    <div class="alert alert-info" role="alert" style="margin-bottom:0px;"
+                        attrs="{'invisible': ['|', ('edi_error_count', '=', 0), ('edi_blocking_level', '!=', 'info')]}">
+                        <div class="o_row">
+                            <field name="edi_error_message" />
+                            <button name="%(account_edi.action_open_edi_documents)d" string="⇒ See errors" type="action" class="oe_link" attrs="{'invisible': [('edi_error_count', '=', 1)]}" /> 
+                        </div>
                     </div>
                 </xpath>
                 <xpath expr="//field[@name='journal_id']" position="after">
