--- addons/account_edi_extended/i18n/account_edi_extended.pot
@@ -1,151 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* account_edi_extended
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 14.0+e\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2021-01-05 11:07+0000\n"
-"PO-Revision-Date: 2021-01-05 11:07+0000\n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: account_edi_extended
-#: code:addons/account_edi_extended/models/account_move.py:0
-#, python-format
-msgid " Electronic invoicing error(s)"
-msgstr ""
-
-#. module: account_edi_extended
-#: code:addons/account_edi_extended/models/account_move.py:0
-#, python-format
-msgid " Electronic invoicing info(s)"
-msgstr ""
-
-#. module: account_edi_extended
-#: code:addons/account_edi_extended/models/account_move.py:0
-#, python-format
-msgid " Electronic invoicing warning(s)"
-msgstr ""
-
-#. module: account_edi_extended
-#: code:addons/account_edi_extended/models/account_move.py:0
-#, python-format
-msgid "A request for cancellation of the EDI has been called off."
-msgstr ""
-
-#. module: account_edi_extended
-#: model:ir.model.fields,field_description:account_edi_extended.field_account_edi_document__blocking_level
-msgid "Blocking Level"
-msgstr ""
-
-#. module: account_edi_extended
-#: model:ir.model.fields,help:account_edi_extended.field_account_edi_document__blocking_level
-msgid ""
-"Blocks the document current operation depending on the error severity :\n"
-"  * Info: the document is not blocked and everything is working as it should.\n"
-"  * Warning : there is an error that doesn't prevent the current Electronic Invoicing operation to succeed.\n"
-"  * Error : there is an error that blocks the current Electronic Invoicing operation."
-msgstr ""
-
-#. module: account_edi_extended
-#: model_terms:ir.ui.view,arch_db:account_edi_extended.view_move_form_inherit
-msgid "Call off EDI Cancellation"
-msgstr ""
-
-#. module: account_edi_extended
-#: model:ir.model.fields,field_description:account_edi_extended.field_account_edi_document__display_name
-#: model:ir.model.fields,field_description:account_edi_extended.field_account_move__display_name
-#: model:ir.model.fields,field_description:account_edi_extended.field_account_payment__display_name
-msgid "Display Name"
-msgstr ""
-
-#. module: account_edi_extended
-#: model:ir.model.fields,field_description:account_edi_extended.field_account_bank_statement_line__edi_blocking_level
-#: model:ir.model.fields,field_description:account_edi_extended.field_account_move__edi_blocking_level
-#: model:ir.model.fields,field_description:account_edi_extended.field_account_payment__edi_blocking_level
-msgid "Edi Blocking Level"
-msgstr ""
-
-#. module: account_edi_extended
-#: model:ir.model.fields,field_description:account_edi_extended.field_account_bank_statement_line__edi_error_message
-#: model:ir.model.fields,field_description:account_edi_extended.field_account_move__edi_error_message
-#: model:ir.model.fields,field_description:account_edi_extended.field_account_payment__edi_error_message
-msgid "Edi Error Message"
-msgstr ""
-
-#. module: account_edi_extended
-#: model:ir.model.fields,field_description:account_edi_extended.field_account_bank_statement_line__edi_show_abandon_cancel_button
-#: model:ir.model.fields,field_description:account_edi_extended.field_account_move__edi_show_abandon_cancel_button
-#: model:ir.model.fields,field_description:account_edi_extended.field_account_payment__edi_show_abandon_cancel_button
-msgid "Edi Show Abandon Cancel Button"
-msgstr ""
-
-#. module: account_edi_extended
-#: model:ir.model,name:account_edi_extended.model_account_edi_document
-msgid "Electronic Document for an account.move"
-msgstr ""
-
-#. module: account_edi_extended
-#: model:ir.model.fields.selection,name:account_edi_extended.selection__account_edi_document__blocking_level__error
-#: model:ir.model.fields.selection,name:account_edi_extended.selection__account_move__edi_blocking_level__error
-msgid "Error"
-msgstr ""
-
-#. module: account_edi_extended
-#: model:ir.model.fields,field_description:account_edi_extended.field_account_edi_document__id
-#: model:ir.model.fields,field_description:account_edi_extended.field_account_move__id
-#: model:ir.model.fields,field_description:account_edi_extended.field_account_payment__id
-msgid "ID"
-msgstr ""
-
-#. module: account_edi_extended
-#: model:ir.model.fields.selection,name:account_edi_extended.selection__account_edi_document__blocking_level__info
-#: model:ir.model.fields.selection,name:account_edi_extended.selection__account_move__edi_blocking_level__info
-msgid "Info"
-msgstr ""
-
-#. module: account_edi_extended
-#: code:addons/account_edi_extended/models/account_edi_document.py:0
-#, python-format
-msgid "Invalid configuration:"
-msgstr ""
-
-#. module: account_edi_extended
-#: model:ir.model,name:account_edi_extended.model_account_move
-msgid "Journal Entry"
-msgstr ""
-
-#. module: account_edi_extended
-#: model:ir.model.fields,field_description:account_edi_extended.field_account_edi_document____last_update
-#: model:ir.model.fields,field_description:account_edi_extended.field_account_move____last_update
-#: model:ir.model.fields,field_description:account_edi_extended.field_account_payment____last_update
-msgid "Last Modified on"
-msgstr ""
-
-#. module: account_edi_extended
-#: model:ir.model,name:account_edi_extended.model_account_payment
-msgid "Payments"
-msgstr ""
-
-#. module: account_edi_extended
-#: model_terms:ir.ui.view,arch_db:account_edi_extended.view_move_form_inherit
-#: model_terms:ir.ui.view,arch_db:account_edi_extended.view_payment_form_inherit
-msgid "Retry"
-msgstr ""
-
-#. module: account_edi_extended
-#: model:ir.model.fields.selection,name:account_edi_extended.selection__account_edi_document__blocking_level__warning
-#: model:ir.model.fields.selection,name:account_edi_extended.selection__account_move__edi_blocking_level__warning
-msgid "Warning"
-msgstr ""
-
-#. module: account_edi_extended
-#: model_terms:ir.ui.view,arch_db:account_edi_extended.view_move_form_inherit
-msgid "⇒ See errors"
-msgstr ""
