--- addons/account_edi_extended/models/__init__.py
@@ -1,3 +1,0 @@
-from . import account_edi_document
-from . import account_move
-from . import account_payment
