--- addons/account_edi_extended/views/account_payment_views.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <data>
-        <record id="view_payment_form_inherit" model="ir.ui.view">
-            <field name="name">account.payment.form.inherit</field>
-            <field name="model">account.payment</field>
-            <field name="inherit_id" ref="account_edi.view_payment_form_inherit" />
-            <field name="arch" type="xml">
-                <xpath expr="//button[@name='%(account_edi.action_open_payment_edi_documents)d']" position="after">
-                    <button name="action_retry_edi_documents_error" type="object" class="oe_link oe_inline" string="Retry" />
-                </xpath>
-            </field>
-        </record>
-    </data>
-</odoo>
